optimize auto-configuration
