+++ seata-spring-boot-starter/src/main/java/io/seata/spring/boot/autoconfigure/SeataAutoConfiguration.java
@@ -16,37 +16,7 @@
 package io.seata.spring.boot.autoconfigure;
 
 import io.seata.spring.annotation.GlobalTransactionScanner;
-import io.seata.spring.annotation.datasource.SeataAutoDataSourceProxyCreator;
-import io.seata.spring.annotation.datasource.SeataDataSourceBeanPostProcessor;
 import io.seata.spring.boot.autoconfigure.properties.SeataProperties;
-import io.seata.spring.boot.autoconfigure.properties.client.LockProperties;
-import io.seata.spring.boot.autoconfigure.properties.client.LogProperties;
-import io.seata.spring.boot.autoconfigure.properties.client.RmProperties;
-import io.seata.spring.boot.autoconfigure.properties.client.ServiceProperties;
-import io.seata.spring.boot.autoconfigure.properties.client.ShutdownProperties;
-import io.seata.spring.boot.autoconfigure.properties.client.ThreadFactoryProperties;
-import io.seata.spring.boot.autoconfigure.properties.client.TmProperties;
-import io.seata.spring.boot.autoconfigure.properties.client.TransportProperties;
-import io.seata.spring.boot.autoconfigure.properties.client.UndoCompressProperties;
-import io.seata.spring.boot.autoconfigure.properties.client.UndoProperties;
-import io.seata.spring.boot.autoconfigure.properties.config.ConfigApolloProperties;
-import io.seata.spring.boot.autoconfigure.properties.config.ConfigConsulProperties;
-import io.seata.spring.boot.autoconfigure.properties.config.ConfigCustomProperties;
-import io.seata.spring.boot.autoconfigure.properties.config.ConfigEtcd3Properties;
-import io.seata.spring.boot.autoconfigure.properties.config.ConfigFileProperties;
-import io.seata.spring.boot.autoconfigure.properties.config.ConfigNacosProperties;
-import io.seata.spring.boot.autoconfigure.properties.config.ConfigProperties;
-import io.seata.spring.boot.autoconfigure.properties.config.ConfigZooKeeperProperties;
-import io.seata.spring.boot.autoconfigure.properties.registry.RegistryConsulProperties;
-import io.seata.spring.boot.autoconfigure.properties.registry.RegistryCustomProperties;
-import io.seata.spring.boot.autoconfigure.properties.registry.RegistryEtcd3Properties;
-import io.seata.spring.boot.autoconfigure.properties.registry.RegistryEurekaProperties;
-import io.seata.spring.boot.autoconfigure.properties.registry.RegistryNacosProperties;
-import io.seata.spring.boot.autoconfigure.properties.registry.RegistryProperties;
-import io.seata.spring.boot.autoconfigure.properties.registry.RegistryRedisProperties;
-import io.seata.spring.boot.autoconfigure.properties.registry.RegistrySofaProperties;
-import io.seata.spring.boot.autoconfigure.properties.registry.RegistryZooKeeperProperties;
-import io.seata.spring.boot.autoconfigure.provider.SpringApplicationContextProvider;
 import io.seata.tm.api.DefaultFailureHandlerImpl;
 import io.seata.tm.api.FailureHandler;
 import org.slf4j.Logger;
@@ -54,95 +24,22 @@ import org.slf4j.LoggerFactory;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.DependsOn;
 
 import static io.seata.common.Constants.BEAN_NAME_FAILURE_HANDLER;
 import static io.seata.common.Constants.BEAN_NAME_SPRING_APPLICATION_CONTEXT_PROVIDER;
-import static io.seata.spring.annotation.datasource.AutoDataSourceProxyRegistrar.BEAN_NAME_SEATA_AUTO_DATA_SOURCE_PROXY_CREATOR;
-import static io.seata.spring.annotation.datasource.AutoDataSourceProxyRegistrar.BEAN_NAME_SEATA_DATA_SOURCE_BEAN_POST_PROCESSOR;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.CLIENT_RM_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.CLIENT_TM_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.COMPRESS_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.CONFIG_APOLLO_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.CONFIG_CONSUL_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.CONFIG_CUSTOM_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.CONFIG_ETCD3_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.CONFIG_FILE_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.CONFIG_NACOS_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.CONFIG_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.CONFIG_ZK_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.LOCK_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.LOG_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.PROPERTY_BEAN_MAP;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.REGISTRY_CONSUL_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.REGISTRY_CUSTOM_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.REGISTRY_ETCD3_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.REGISTRY_EUREKA_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.REGISTRY_NACOS_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.REGISTRY_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.REGISTRY_REDIS_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.REGISTRY_SOFA_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.REGISTRY_ZK_PREFIX;
 import static io.seata.spring.boot.autoconfigure.StarterConstants.SEATA_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.SERVICE_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.SHUTDOWN_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.THREAD_FACTORY_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.TRANSPORT_PREFIX;
-import static io.seata.spring.boot.autoconfigure.StarterConstants.UNDO_PREFIX;
 
 /**
+ * The type Seata auto configuration
+ *
  * @author xingfudeshi@gmail.com
  */
-@ComponentScan(basePackages = "io.seata.spring.boot.autoconfigure.properties")
 @ConditionalOnProperty(prefix = SEATA_PREFIX, name = "enabled", havingValue = "true", matchIfMissing = true)
 @Configuration
 public class SeataAutoConfiguration {
     private static final Logger LOGGER = LoggerFactory.getLogger(SeataAutoConfiguration.class);
-
-    static {
-
-        PROPERTY_BEAN_MAP.put(SEATA_PREFIX, SeataProperties.class);
-
-        PROPERTY_BEAN_MAP.put(CLIENT_RM_PREFIX, RmProperties.class);
-        PROPERTY_BEAN_MAP.put(CLIENT_TM_PREFIX, TmProperties.class);
-        PROPERTY_BEAN_MAP.put(LOCK_PREFIX, LockProperties.class);
-        PROPERTY_BEAN_MAP.put(SERVICE_PREFIX, ServiceProperties.class);
-        PROPERTY_BEAN_MAP.put(SHUTDOWN_PREFIX, ShutdownProperties.class);
-        PROPERTY_BEAN_MAP.put(THREAD_FACTORY_PREFIX, ThreadFactoryProperties.class);
-        PROPERTY_BEAN_MAP.put(UNDO_PREFIX, UndoProperties.class);
-        PROPERTY_BEAN_MAP.put(COMPRESS_PREFIX, UndoCompressProperties.class);
-        PROPERTY_BEAN_MAP.put(LOG_PREFIX, LogProperties.class);
-        PROPERTY_BEAN_MAP.put(TRANSPORT_PREFIX, TransportProperties.class);
-        PROPERTY_BEAN_MAP.put(CONFIG_PREFIX, ConfigProperties.class);
-        PROPERTY_BEAN_MAP.put(CONFIG_FILE_PREFIX, ConfigFileProperties.class);
-        PROPERTY_BEAN_MAP.put(REGISTRY_PREFIX, RegistryProperties.class);
-
-        PROPERTY_BEAN_MAP.put(CONFIG_NACOS_PREFIX, ConfigNacosProperties.class);
-        PROPERTY_BEAN_MAP.put(CONFIG_CONSUL_PREFIX, ConfigConsulProperties.class);
-        PROPERTY_BEAN_MAP.put(CONFIG_ZK_PREFIX, ConfigZooKeeperProperties.class);
-        PROPERTY_BEAN_MAP.put(CONFIG_APOLLO_PREFIX, ConfigApolloProperties.class);
-        PROPERTY_BEAN_MAP.put(CONFIG_ETCD3_PREFIX, ConfigEtcd3Properties.class);
-        PROPERTY_BEAN_MAP.put(CONFIG_CUSTOM_PREFIX, ConfigCustomProperties.class);
-
-        PROPERTY_BEAN_MAP.put(REGISTRY_CONSUL_PREFIX, RegistryConsulProperties.class);
-        PROPERTY_BEAN_MAP.put(REGISTRY_ETCD3_PREFIX, RegistryEtcd3Properties.class);
-        PROPERTY_BEAN_MAP.put(REGISTRY_EUREKA_PREFIX, RegistryEurekaProperties.class);
-        PROPERTY_BEAN_MAP.put(REGISTRY_NACOS_PREFIX, RegistryNacosProperties.class);
-        PROPERTY_BEAN_MAP.put(REGISTRY_REDIS_PREFIX, RegistryRedisProperties.class);
-        PROPERTY_BEAN_MAP.put(REGISTRY_SOFA_PREFIX, RegistrySofaProperties.class);
-        PROPERTY_BEAN_MAP.put(REGISTRY_ZK_PREFIX, RegistryZooKeeperProperties.class);
-        PROPERTY_BEAN_MAP.put(REGISTRY_CUSTOM_PREFIX, RegistryCustomProperties.class);
-
-    }
-
-
-    @Bean(BEAN_NAME_SPRING_APPLICATION_CONTEXT_PROVIDER)
-    @ConditionalOnMissingBean(name = {BEAN_NAME_SPRING_APPLICATION_CONTEXT_PROVIDER})
-    public SpringApplicationContextProvider springApplicationContextProvider() {
-        return new SpringApplicationContextProvider();
-    }
 
     @Bean(BEAN_NAME_FAILURE_HANDLER)
     @ConditionalOnMissingBean(FailureHandler.class)
@@ -151,34 +48,7 @@ public class SeataAutoConfiguration {
     }
 
     @Bean
-    @DependsOn({"seataProperties",
-        "rmProperties",
-        "tmProperties",
-        "lockProperties",
-        "serviceProperties",
-        "shutdownProperties",
-        "threadFactoryProperties",
-        "undoProperties",
-        "undoCompressProperties",
-        "logProperties",
-        "transportProperties",
-        "configProperties",
-        "configFileProperties",
-        "registryProperties",
-        "configNacosProperties",
-        "configConsulProperties",
-        "configZooKeeperProperties",
-        "configApolloProperties",
-        "configEtcd3Properties",
-        "configCustomProperties",
-        "registryConsulProperties",
-        "registryEtcd3Properties",
-        "registryEurekaProperties",
-        "registryNacosProperties",
-        "registryRedisProperties",
-        "registrySofaProperties",
-        "registryZooKeeperProperties",
-        "registryCustomProperties", BEAN_NAME_SPRING_APPLICATION_CONTEXT_PROVIDER, BEAN_NAME_FAILURE_HANDLER})
+    @DependsOn({BEAN_NAME_SPRING_APPLICATION_CONTEXT_PROVIDER, BEAN_NAME_FAILURE_HANDLER})
     @ConditionalOnMissingBean(GlobalTransactionScanner.class)
     public GlobalTransactionScanner globalTransactionScanner(SeataProperties seataProperties, FailureHandler failureHandler) {
         if (LOGGER.isInfoEnabled()) {
@@ -188,30 +58,4 @@ public class SeataAutoConfiguration {
     }
 
 
-    /**
-     * The data source configuration.
-     */
-    @Configuration
-    @ConditionalOnProperty(prefix = SEATA_PREFIX, name = {"enableAutoDataSourceProxy", "enable-auto-data-source-proxy"}, havingValue = "true", matchIfMissing = true)
-    static class SeataDataSourceConfiguration {
-
-        /**
-         * The bean seataDataSourceBeanPostProcessor.
-         */
-        @Bean(BEAN_NAME_SEATA_DATA_SOURCE_BEAN_POST_PROCESSOR)
-        @ConditionalOnMissingBean(SeataDataSourceBeanPostProcessor.class)
-        public SeataDataSourceBeanPostProcessor seataDataSourceBeanPostProcessor(SeataProperties seataProperties) {
-            return new SeataDataSourceBeanPostProcessor(seataProperties.getExcludesForAutoProxying(), seataProperties.getDataSourceProxyMode());
-        }
-
-        /**
-         * The bean seataAutoDataSourceProxyCreator.
-         */
-        @Bean(BEAN_NAME_SEATA_AUTO_DATA_SOURCE_PROXY_CREATOR)
-        @ConditionalOnMissingBean(SeataAutoDataSourceProxyCreator.class)
-        public SeataAutoDataSourceProxyCreator seataAutoDataSourceProxyCreator(SeataProperties seataProperties) {
-            return new SeataAutoDataSourceProxyCreator(seataProperties.isUseJdkProxy(),
-                seataProperties.getExcludesForAutoProxying(), seataProperties.getDataSourceProxyMode());
-        }
-    }
 }
