+++ seata-spring-boot-starter/src/main/java/io/seata/spring/boot/autoconfigure/SeataDataSourceConfiguration.java
@@ -1,0 +1,42 @@
+package io.seata.spring.boot.autoconfigure;
+
+import io.seata.spring.annotation.datasource.SeataAutoDataSourceProxyCreator;
+import io.seata.spring.annotation.datasource.SeataDataSourceBeanPostProcessor;
+import io.seata.spring.boot.autoconfigure.properties.SeataProperties;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
+import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+
+import static io.seata.spring.annotation.datasource.AutoDataSourceProxyRegistrar.BEAN_NAME_SEATA_AUTO_DATA_SOURCE_PROXY_CREATOR;
+import static io.seata.spring.annotation.datasource.AutoDataSourceProxyRegistrar.BEAN_NAME_SEATA_DATA_SOURCE_BEAN_POST_PROCESSOR;
+import static io.seata.spring.boot.autoconfigure.StarterConstants.SEATA_PREFIX;
+
+/**
+ * The type Seata data source configuration.
+ *
+ * @author xingfudeshi@gmail.com
+ */
+@Configuration
+@ConditionalOnProperty(prefix = SEATA_PREFIX, name = {"enableAutoDataSourceProxy", "enable-auto-data-source-proxy"}, havingValue = "true", matchIfMissing = true)
+public class SeataDataSourceConfiguration {
+
+    /**
+     * The bean seataDataSourceBeanPostProcessor.
+     */
+    @Bean(BEAN_NAME_SEATA_DATA_SOURCE_BEAN_POST_PROCESSOR)
+    @ConditionalOnMissingBean(SeataDataSourceBeanPostProcessor.class)
+    public SeataDataSourceBeanPostProcessor seataDataSourceBeanPostProcessor(SeataProperties seataProperties) {
+        return new SeataDataSourceBeanPostProcessor(seataProperties.getExcludesForAutoProxying(), seataProperties.getDataSourceProxyMode());
+    }
+
+    /**
+     * The bean seataAutoDataSourceProxyCreator.
+     */
+    @Bean(BEAN_NAME_SEATA_AUTO_DATA_SOURCE_PROXY_CREATOR)
+    @ConditionalOnMissingBean(SeataAutoDataSourceProxyCreator.class)
+    public SeataAutoDataSourceProxyCreator seataAutoDataSourceProxyCreator(SeataProperties seataProperties) {
+        return new SeataAutoDataSourceProxyCreator(seataProperties.isUseJdkProxy(),
+            seataProperties.getExcludesForAutoProxying(), seataProperties.getDataSourceProxyMode());
+    }
+}
