+++ seata-spring-boot-starter/src/main/resources/META-INF/spring.factories
@@ -1,4 +1,6 @@
 # Auto Configure
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+io.seata.spring.boot.autoconfigure.SeataPropertiesConfiguration,\
+io.seata.spring.boot.autoconfigure.SeataDataSourceConfiguration,\
 io.seata.spring.boot.autoconfigure.SeataAutoConfiguration,\
 io.seata.spring.boot.autoconfigure.HttpAutoConfiguration
